fixing docker flow

+++ internal/dto/products_dto.go
@@ -1,14 +1,14 @@
 package dto
 
 type ProductInputDTO struct {
-	ID          int     `json:"id"`
+	ID          int64   `json:"id"`
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
 	Price       float64 `json:"price"`
 }
 
 type ProductOutputDTO struct {
-	ID          int     `json:"id"`
+	ID          int64   `json:"id"`
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
 	Price       float64 `json:"price"`
